added save and delete alarm from database

+++ app/src/main/java/ru/mirea/ivashechkinav/alarmclock/data/room/models/AlarmRoom.kt
@@ -19,10 +19,10 @@ data class AlarmRoom(
     override val daysOfWeek: EnumSet<DaysOfWeek>,
 ) : Alarm {
     @PrimaryKey(autoGenerate = true)
-    var id: Int? = null
+    var id: Long? = null
 
     override val requestCode: Int
-        get() = id!!
+        get() = id!!.toInt()
 
     constructor(alarm: Alarm) : this(
         name = alarm.name,
